portal/requirements.txt
ansible==7.2.0
ansible-core==2.14.2
ansiwrap==0.8.4
art==5.9
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.1
distlib==0.3.6
Fabric3==1.14.post1
filelock==3.9.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
paramiko==2.12.0
platformdirs==2.6.2
pycparser==2.21
PyNaCl==1.5.0
python-vagrant==1.0.0
PyYAML==6.0
resolvelib==0.8.1
six==1.16.0
textwrap3==0.9.2
virtualenv==20.18.0
